app/requirements.txt.txt
streamlit
pyAgrum
pandas
Pillow